Web Inspector: Attempting to select records in the bottom 16px of the timeline overview graph fails https://bugs.webkit.org/show_bug.cgi?id=195318
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2019-03-05
Reviewed by Devin Rousso.

The bottom few pixels of the overview are reserved for the scroll-container
which may show a scrollbar if the overview needs to scroll. When the scrollbars
are not visible we can allow pointer events to click through the invisible
scroll-container and allow record selection.

* UserInterface/Views/TimelineOverview.css:
(.timeline-overview:not(.has-scrollbar) > .scroll-container):
* UserInterface/Views/TimelineOverview.js:
(WI.TimelineOverview.prototype._handleScrollEvent):
(WI.TimelineOverview.prototype._handleWheelEvent):
(WI.TimelineOverview._handleGestureStart):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@242510 268f45cc-cd09-0410-ab3c-d52691b4dbfc