[CMake] Prevent potential warning when determining $architecture https://bugs.webkit.org/show_bug.cgi?id=195891
Patch by Xan Lopez <xan@igalia.com> on 2019-03-18
Reviewed by Michael Catanzaro.

It's possible for generateBuildSystemForCMakeProject to be called
before determineArchitecture() is ever called. This will make the
first eq check with $architecture to be bogus (and print a
warning). Change that to an eq check with architecture(), which
implicitly calls determineArchitecture() if needed, and change all
the other similar cases in the method in the same way to prevent
future bugs like this one.

* Scripts/webkitdirs.pm:
(generateBuildSystemFromCMakeProject):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243083 268f45cc-cd09-0410-ab3c-d52691b4dbfc