REGRESSION (r243094): internal media test fairplay-hls-error.html is failing https://bugs.webkit.org/show_bug.cgi?id=195954 <rdar://problem/49005981>
Reviewed by Geoffrey Garen.

TestController::resetStateToConsistentValues() tries to send a "reset" message to its
injected bundle. Part of the "reset" message sets the external hosts that we're allowed
to load. If there is no WebProcess yet when WKPagePostMessageToInjectedBundle() is called,
then this message does not get sent and the allowed hosts do not get set, causing this
test failure.

To address the issue, make sure we launch the initial process if necessary when
WebPageProxy::postMessageToInjectedBundle() is called, in order to maintain backward
compatibility.

* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::postMessageToInjectedBundle):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243156 268f45cc-cd09-0410-ab3c-d52691b4dbfc