[GStreamer][WebRTC]: Refactor to make the encoding happen synchronously https://bugs.webkit.org/show_bug.cgi?id=195674
This simplifies the code and make sure that everything is called from the right
thread. It also makes it so we follow the same way of doing things as other libwebrtc
encoders.

Also "s/codecSpecifiInfos/codecSpecificInfos/"

Patch by Thibault Saunier <tsaunier@igalia.com> on 2019-03-20
Reviewed by Philippe Normand.

No new test as this is just a refactoring

* platform/mediastream/libwebrtc/GStreamerVideoEncoderFactory.cpp:
(WebCore::GStreamerVideoEncoder::InitEncode):
(WebCore::GStreamerVideoEncoder::returnFromFlowReturn):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243196 268f45cc-cd09-0410-ab3c-d52691b4dbfc