storage/indexeddb/closed-cursor-private.html is crashing on debug https://bugs.webkit.org/show_bug.cgi?id=196101
Reviewed by Alex Christensen.

In case of immediate close for user delete, the transaction might be deleted.
If the request for space is pending, it will be executed with an error as parameter.
In that case, the call to didCommitTransaction will not find any existing transaction.
Speculative fix, test should no longer crash.

* Modules/indexeddb/server/UniqueIDBDatabaseConnection.cpp:
(WebCore::IDBServer::UniqueIDBDatabaseConnection::didCommitTransaction):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243401 268f45cc-cd09-0410-ab3c-d52691b4dbfc