REGRESSION(r243512): Change locale of test in intl-datetimeformat.js https://bugs.webkit.org/show_bug.cgi?id=196395
r243512 modifes Intl.DateTimeFormat to obey 2-digit hour. A test was
added to test 2-digit hour in Russian using 12-hour setting as true and
false. In the former case, the expected string appends the string 'AM'
in Russian. This localised string made GTK-based ports fail since
according to Glib the 'am_pm' values in Russian are not localised.

This patch keeps the same test but changes locale to 'en' to prevent
the test from failing.

Reviewed by Michael Catanzaro.

* js/intl-datetimeformat-expected.txt:
* js/script-tests/intl-datetimeformat.js:
* platform/gtk/js/intl-datetimeformat-expected.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243740 268f45cc-cd09-0410-ab3c-d52691b4dbfc