[JSC] to_index_string should not assume incoming value is Uint32 https://bugs.webkit.org/show_bug.cgi?id=196713
Reviewed by Saam Barati.

JSTests:

* stress/to-index-string-should-not-assume-incoming-value-is-uint32.js: Added.
(foo):

Source/JavaScriptCore:

The slow path of to_index_string assumes that incoming value is Uint32. But we should not have
this assumption since DFG may decide we should have it double format. This patch removes this
assumption, and instead, we should assume that incoming value is AnyInt and the range of this
is within Uint32.

* runtime/CommonSlowPaths.cpp:
(JSC::SLOW_PATH_DECL):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244057 268f45cc-cd09-0410-ab3c-d52691b4dbfc