Web Inspector: Extension scripts with parse errors do not show up in Web Inspector https://bugs.webkit.org/show_bug.cgi?id=196996 <rdar://problem/47054804>
Reviewed by Devin Rousso.

* UserInterface/Controllers/DebuggerManager.js:
(WI.DebuggerManager.prototype.scriptDidFail):
Create a LocalScript that can be linked to but ignore it
if we already have a Resource for the URL.

* UserInterface/Protocol/DebuggerObserver.js:
(WI.DebuggerObserver.prototype.scriptFailedToParse):
Let DebuggerManager potentially create a script for this.

* UserInterface/Models/Script.js:
(WI.Script):
Local scripts provide a null id and id is not required below.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244398 268f45cc-cd09-0410-ab3c-d52691b4dbfc