Remove DeprecatedOptional https://bugs.webkit.org/show_bug.cgi?id=197161
Patch by Alex Christensen <achristensen@webkit.org> on 2019-04-23
Reviewed by Darin Adler.

Source/JavaScriptCore:

* inspector/InspectorBackendDispatcher.cpp:
* inspector/InspectorBackendDispatcher.h:

Source/WTF:

This was added in r209326 to be compatible with a shipping version of Safari.
We have released several versions of Safari since then, so do what the comments say and remove it.
The existence of this std::optional makes migrating to C++17 harder, and there's no reason to keep it.

* WTF.xcodeproj/project.pbxproj:
* wtf/CMakeLists.txt:
* wtf/DeprecatedOptional.h: Removed.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244558 268f45cc-cd09-0410-ab3c-d52691b4dbfc