[iOS] Calling preventDefault() when handling a pointerdown event should not prevent panning, zooming or click event dispatch https://bugs.webkit.org/show_bug.cgi?id=195839 <rdar://problem/48946154>
Reviewed by Brent Fulgham.

Source/WebCore:

Tests: pointerevents/ios/pointer-events-prevent-default-allows-click-event.html
       pointerevents/ios/pointer-events-prevent-default-allows-scrolling.html

The Pointer Events specification defines that the default action of any and all pointer events MUST NOT
be a manipulation of the viewport (e.g. panning or zooming). In practice, this means that calling
preventDefault() while handling a Pointer Event has no effect on the inner workings of the user agent,
so we change the method signature of PointerCaptureController::dispatchEventForTouchAtIndex() to return
void since we don't need to know whether preventDefault() was called.

https://www.w3.org/TR/pointerevents/#declaring-candidate-regions-for-default-touch-behaviors

* page/PointerCaptureController.cpp:
(WebCore::PointerCaptureController::dispatchEventForTouchAtIndex):
* page/PointerCaptureController.h:

LayoutTests:

Modify the existing test to check that we *have* scrolled after performing a swipe gesture and calling
preventDefault() while handling the pointerdown handling. We also add a new test that checks that the
click event is indeed dispatched when tapping on an element and calling preventDefault() while handling
the pointerdown event.

* pointerevents/ios/pointer-events-prevent-default-allows-click-event-expected.txt: Added.
* pointerevents/ios/pointer-events-prevent-default-allows-click-event.html: Added.
* pointerevents/ios/pointer-events-prevent-default-allows-scrolling-expected.txt: Added.
* pointerevents/ios/pointer-events-prevent-default-allows-scrolling.html: Renamed from LayoutTests/pointerevents/ios/pointer-events-prevent-default.html.
* pointerevents/ios/pointer-events-prevent-default-expected.txt: Removed.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244598 268f45cc-cd09-0410-ab3c-d52691b4dbfc