[CMake] Add WEBKIT_EXECUTABLE macro https://bugs.webkit.org/show_bug.cgi?id=197206
Reviewed by Konstantin Tokarev.

.:

Add WEBKIT_EXECUTABLE macro for creating executable targets. This wraps the creation of
executable targets using the same conventions of WEBKIT_FRAMEWORK.

Common code from the two macros were moved to WEBKIT_TARGET which is meant as an internal
macro.

Add WEBKIT_WRAP_EXECUTABLE macro to support Windows targets which use a DLLLauncherMain to
launch the executable.

* Source/cmake/WebKitMacros.cmake:

Source/JavaScriptCore:

Migrate to WEBKIT_EXECUTABLE for the jsc and test targets.

* b3/air/testair.cpp:
* b3/testb3.cpp:
* dfg/testdfg.cpp:
* shell/CMakeLists.txt:
* shell/PlatformGTK.cmake:
* shell/PlatformJSCOnly.cmake: Removed.
* shell/PlatformMac.cmake:
* shell/PlatformPlayStation.cmake:
* shell/PlatformWPE.cmake:
* shell/PlatformWin.cmake:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244712 268f45cc-cd09-0410-ab3c-d52691b4dbfc