Enable new build rule for post-processing headers when using XCBuild https://bugs.webkit.org/show_bug.cgi?id=197340 <rdar://problem/50226685>
Reviewed by Brent Fulgham.

In Bug 197116, we conditionally disabled the old method for
post-processing header files when we are using the new XCBuild build
system. This check-in conditionally enables the new post-processing
facility. Note that the old system is disabled and the new system
enabled only when the USE_NEW_BUILD_SYSTEM environment variable is set
to YES.

Source/JavaScriptCore:

* Configurations/JavaScriptCore.xcconfig:

Source/WebKit:

* Configurations/WebKit.xcconfig:

Source/WebKitLegacy/mac:

* Configurations/WebKitLegacy.xcconfig:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244715 268f45cc-cd09-0410-ab3c-d52691b4dbfc