iOS: Selection is dismissed even if click is preventDefault()'d https://bugs.webkit.org/show_bug.cgi?id=197686 <rdar://problem/49398824>
Reviewed by Wenson Hsieh.

Source/WebKit:

We currently unconditionally dismiss the selection on any tap; however
if a site preventDefault()s on click, we shouldn't perform the default
action of dismissing the selection.

Instead of clearing the selection in the UI process, clear it in the
Web content process if we don't dispatch a synthetic click; the normal
WebCore machinery will handle it in the case that we do.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView _singleTapRecognized:]):
* WebProcess/WebPage/WebPage.h:
* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::commitPotentialTapFailed):
(WebKit::WebPage::selectWithGesture):
(WebKit::WebPage::clearSelection):
(WebKit::WebPage::selectTextWithGranularityAtPoint):

LayoutTests:

* editing/selection/ios/clear-selection-after-tapping-on-element-with-no-click-handler-expected.txt: Added.
* editing/selection/ios/clear-selection-after-tapping-on-element-with-no-click-handler.html: Added.
* editing/selection/ios/persist-selection-after-tapping-on-element-with-click-handler-expected.txt: Added.
* editing/selection/ios/persist-selection-after-tapping-on-element-with-click-handler.html: Added.
New tests.

* platform/ios/editing/deleting/smart-delete-paragraph-003-expected.txt:
* platform/ios/editing/pasteboard/smart-paste-paragraph-003-expected.txt:
Rebaseline since we changed the timing of editing callbacks by changing where the selection happens.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245067 268f45cc-cd09-0410-ab3c-d52691b4dbfc