[iOS] The shouldIgnoreMetaViewport test option incorrectly persists across layout tests https://bugs.webkit.org/show_bug.cgi?id=198135 <rdar://problem/51001306>
Reviewed by Alexey Proskuryakov.

Tools:

When shouldIgnoreMetaViewport=true is specified on a test, if the following test does not use a flexible
viewport, we will continue ignoring the meta viewport, since we'll bail before we attempt to reset
WKPreferences' _shouldIgnoreMetaViewport property to NO.

We fix this by simply allowing shouldIgnoreMetaViewport to take precedence over useFlexibleViewport, such that
-[WKPreferences _shouldIgnoreMetaViewport] will always be updated to its intended value between tests.

* WebKitTestRunner/ios/TestControllerIOS.mm:
(WTR::TestController::platformConfigureViewForTest):

LayoutTests:

Attempt to fix a handful of viewport tests that currently try to wait for the next visible content rect update,
but end up timing out in the process. Instead, only wait for the presentation update, which is guaranteed to
happen.

* fast/viewport/ios/constant-width-viewport-after-changing-view-scale.html:
* fast/viewport/ios/device-width-viewport-after-changing-view-scale.html:
* fast/viewport/ios/initial-scale-after-changing-view-scale.html:
* fast/viewport/ios/minimum-scale-after-changing-view-scale.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245663 268f45cc-cd09-0410-ab3c-d52691b4dbfc