Web Inspector: Overlay: rulers should switch sides if they intersect the highlighted node(s) so they don't obstruct any content https://bugs.webkit.org/show_bug.cgi?id=198165
Reviewed by Timothy Hatcher.

If the highlighted node is against the top edge of the screen, the top ruler should shift to
the bottom, unless the highlighted node is also against the bottom edge of the screen.

If the highlighted node is against the left edge of the screen, the left ruler should shift
to the right, unless the highlighted node is also against the right edge of the screen.

This way, unless the node is very wide/tall, the rulers won't be drawn on top of anything
being highlighted.

* inspector/InspectorOverlay.h:
* inspector/InspectorOverlay.cpp:
(WebCore::InspectorOverlay::paint):
(WebCore::InspectorOverlay::drawNodeHighlight):
(WebCore::InspectorOverlay::drawQuadHighlight):
(WebCore::InspectorOverlay::drawBounds):
(WebCore::InspectorOverlay::drawRulers):
Drive-by: create an alias for the type (`FloatRect`) used when calculating the bounds of
          everything that's highlighted.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245728 268f45cc-cd09-0410-ab3c-d52691b4dbfc