Asssertion failure in dispatchSubtreeModifiedEvent due to TextFieldInputType updating UA shadow tree inside Element::removedFromAncestor https://bugs.webkit.org/show_bug.cgi?id=198216
Reviewed by Brent Fulgham.

Source/WebCore:

The bug was caused by ListAttributeTargetObserver::idTargetChanged() updating the shadow tree of an input element
within Element::removedFromAncestor via TextFieldInputType::createDataListDropdownIndicator(). Fixed it by
supressing the assertions with ScriptDisallowedScope::EventAllowedScope since it's always safe to update
UA shadow trees of input elements as it's not exposed to author scripts.

Avoiding the creation of dropdown indicator in this particular scenario is a lot more involved and it's not
particularly correct because there could be another datalist element which matches the ID specified in list
content attribute after the removal of the old datalist element.

Test: fast/forms/datalist/datalist-removal-assertion.html

* html/TextFieldInputType.cpp:
(WebCore::TextFieldInputType::createDataListDropdownIndicator):
(WebCore::TextFieldInputType::createContainer):
* html/shadow/DataListButtonElement.cpp:
(WebCore::DataListButtonElement::DataListButtonElement):

LayoutTests:

Added a regression test.

* fast/forms/datalist/datalist-removal-assertion-expected.txt: Added.
* fast/forms/datalist/datalist-removal-assertion.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245746 268f45cc-cd09-0410-ab3c-d52691b4dbfc