Make computed width of non-replaced inline return computed style. https://bugs.webkit.org/show_bug.cgi?id=197814
Reviewed by Antti Koivisto.

Currently, Computed width of non-replaced inline incorrectly returns "auto"
instead of the computed value.
This patch changes the behavior according to
https://drafts.csswg.org/cssom/#resolved-value as below.

'If the property applies to the element or pseudo-element
and the resolved value of the display property is not none or contents,
then the resolved value is the used value.
Otherwise the resolved value is the computed value.'

LayoutTests/imported/w3c:

* web-platform-tests/html/semantics/embedded-content/the-object-element/object-attributes-expected.txt:

Source/WebCore:

* css/CSSComputedStyleDeclaration.cpp:
(WebCore::ComputedStyleExtractor::valueForPropertyinStyle):

LayoutTests:

* fast/css/getComputedStyle/getComputedStyle-resolved-values-expected.txt:
* fast/css/getComputedStyle/getComputedStyle-with-pseudo-element-expected.txt:
* fast/css/getComputedStyle/getComputedStyle-with-pseudo-element.html:
* legacy-animation-engine/fast/css/getComputedStyle/getComputedStyle-with-pseudo-element-expected.txt:
* legacy-animation-engine/fast/css/getComputedStyle/getComputedStyle-with-pseudo-element.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245768 268f45cc-cd09-0410-ab3c-d52691b4dbfc